requirements/development.txt
-r base.txt
django-debug-toolbar==4.2.0
django-extensions==3.2.3
ipython==8.17.2
jupyter==1.0.0
flake8==6.1.0
black==23.11.0
isort==5.12.0
pylint==3.0.3
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
